% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no2.R
\docType{data}
\name{no2}
\alias{no2}
\title{Tibble con i dati di no2 per 32 stazioni della regione VENETO}
\format{
Un tibble con 8 colonne e 23392 osservazioni
}
\usage{
no2
}
\description{
Tibble con i dati di no2 per 32 stazioni della regione VENETO
}
\keyword{datasets}
