% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o3_max_h_d.R
\docType{data}
\name{o3_max_h_d}
\alias{o3_max_h_d}
\title{Tibble con i dati di o3_max_h_d per 23 stazioni della regione VENETO}
\format{
Un tibble con 8 colonne e 60807 osservazioni
}
\usage{
o3_max_h_d
}
\description{
Tibble con i dati di o3_max_h_d per 23 stazioni della regione VENETO
}
\keyword{datasets}
