% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nox.R
\docType{data}
\name{nox}
\alias{nox}
\title{Tibble con i dati di nox per 32 stazioni della regione VENETO}
\format{
Un tibble con 8 colonne e 23392 osservazioni
}
\usage{
nox
}
\description{
Tibble con i dati di nox per 32 stazioni della regione VENETO
}
\keyword{datasets}
